fix: Renomear o controler de enterprise para manter o padrao singular
