+++ src/routes/enterprise.routes.js
@@ -3,10 +3,10 @@ import EnterpriseController from "../controllers/enterprise.controller.js";
 
 const enterpriseRoutes = new Router();
 
-enterpriseRoutes.get("/oppotunity/", EnterpriseController.index);
-enterpriseRoutes.get("/oppotunity/:id", EnterpriseController.show);
-enterpriseRoutes.post("/oppotunity/", EnterpriseController.store);
-enterpriseRoutes.put("/oppotunity/:id", EnterpriseController.update);
-enterpriseRoutes.delete("/oppotunity/:id", EnterpriseController.delete);
+enterpriseRoutes.get("/enterprise/", EnterpriseController.index);
+enterpriseRoutes.get("/enterprise/:id", EnterpriseController.show);
+enterpriseRoutes.post("/enterprise/", EnterpriseController.store);
+enterpriseRoutes.put("/enterprise/:id", EnterpriseController.update);
+enterpriseRoutes.delete("/enterprise/:id", EnterpriseController.delete);
 
 export default enterpriseRoutes;
